G15 support should now do for most people, it shows which weapon is currently selection, if reloading any weapon, charge % for all weapons, and switches to splash screen the moment weapon information is no longer usefull (dead/back to menu)
But i get some weird errors (not related to me), i get a cl: Unknown packet when playing on other people's servers, and after that network traffic seems to stop (have not logged anything, this is all from ingame + console) - simulation for local play still works, but all other worms freeze, the ping is still updated - so i can't get dropped from the server cause of a ping timeout, and no more data is written in the console.
For all i know this happened somewhere between my last commit, and this one.
Pelya does not have this.



git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@2329 34602234-ff1f-0410-a465-ea8f3b77ab7f