Revert "CTF: teams can set up their bases positions before starting the game"
This reverts commit 340f1bc91edb78f59b7f738cae8219207ba895dc.

This is experimental and probably not everybody want that (I don't) and thus make it please an extra branch. Such thing also needs an additional option which enables/disables that. And perhaps it is fully obsolete when we have later a nicer way to set the flag points (for example the hoster can just click there). I am not sure if there should be such workaround way now when it will be different later.