fix for dwarf dumper: handle DW_AT_specification attributes (references to other DIEs) and iterate all child DIEs
It seems that my GCC only sets DW_AT_specification attributes for all C++ functions. After adding that
additional handling for that attributes, I figured out that the DIEs where it was refering have not
been handled at all (i.e. FindChildHandler has never seen them).

The second problem was that these specific DIEs were all childs of other DIEs which were completly
skipped. In most cases it was DW_TAG_class_type. Anyway, to avoid any such problems, I added a dummy
DIEHandler which only wraps FindChildHandler to the RootDIEHandler. With this addition, all DIEs in
the DW_AT_specification attribute were found.

The result of this fix is, that I finally have function names for all my C++ functions. Though, the
functionnames itself are only short versions of the full name. This is probably because the DIE is a
child of another DIE and we have to concatenate the path in the DIE tree from the root to the child.

Also, the crash reporter itself is still not able to find the function symbols from the stacktrace.