fix for netstream: callbacks for replicators only from Net_processOutput
The problem was that in register node, we would have called the replicators. Inside of the replicators, we except though that everything is loaded and the game is running.

In the map init function, this can be a problem, where we haven't loaded everything yet (esp not indexated the particle types) and some of the replicators fail because of this.