changing section for projectile relative speed
Btw., the reason i put that back in is because there simply is *no* reason to not have it.
While it may be a valid argument that this is something the user should not set via the menu,
it still can be nice to have (for example for a mod).

See also my sugestions about game settings presets / moving many options to game mods.