only list human player profiles in netplay for joining a server
The code was already there but commented out. Why? I think it doesn't make any sense for any normal player to join with a bot to a server.

This may only be usefull for debugging. Anyway who really want to do that can do it via console (connect command) or dedicated script.