some different ExactFilenameCache hash function
The old one was ineffective anyway as it was case sensitive and thus the whole purpose (two case insensitive equally strings should have the same hash) was broken.

The new one is case insensitive and based on CRC32.

Not sure though how much improvement (if any) the whole thing brings.