redesigned the game loop. unified some state variables
Basically, we have a fixed 100 FPS inner game loop now. This has certain advantages. Also, we keep a serverFrame variable which is synced so that clients know when we are slow at simulation.

We also have merged many game state variables which have been in the server and client earlier and cleaned that all up.

Note about a small incompatibility change in the dedicated control: getGameState is slightly different now.