Check max weapon selection time when immediate start is on
Max weapon selection time can now be enforced also when immediate
start is enabled. However, the warning message is not displayed
when immediate start is on. This could be fixed, but it would
be annoying for those who are already playing. A better solution
would be to explicitly check whether each player is playing or
still selecting weapons.