Chat annoyance mitigation
The newline tag allows annoying impersonation.
Prevent this by adding "* " immediately after the newline.
This should be done server side so that it doesn't depend on
how the client handles the tag.
Messages containing these tags could be dropped, or the tag could be removed,
but this is a clean and simple way, and should be enough.