Reduced logging
Previously, OLX logged all chat messages to the main log when hosting a server (dedicated or normal).
This cluttered the log and didn't provide very useful information.
Now this logging can be turned off.
NOTE that this has nothing to do with the "Log conversations" option - it goes to a different log!