[dedicated] Find writable path for log file
See #802. The default log file name in dedicated_config doesn't specify a path
so the script location will be used. It might not be writable for normal users.
Another way would be specifying the path explicitly but I think this is consistent
with the other scripts, for example the path for pwn0meter.txt is also searched this way.