[0.58] Update pwn0meter.py (read note before merging) (#820)
* Removed dead master server

* Check max weapon selection time when immediate start is on

Max weapon selection time can now be enforced also when immediate
start is enabled. However, the warning message is not displayed
when immediate start is on. This could be fixed, but it would
be annoying for those who are already playing. A better solution
would be to explicitly check whether each player is playing or
still selecting weapons.

* Bug/glitch fix

Fixed the "empty name join glitch"

* Fixed annoyance

Added nickname length check to the server
- this prevents excessively long nicks
which can cause lots of annoyance. This doesn't
affect normal players as such nicks could be
only created using tricks.

* Fixed annoyance

Same as the previous commit but for "wants to join"
messages - check if the nick is longer than
"legitimately" possible to avoid annoyance.

* Configurable killing/dying spree thresholds

Killing/dying spree threshold values, i.e. the number
of kills/deaths required for each "spree",
can now be configured.

* Name length annoyance fix

Changed the name length check to truncate oversized
names instead of kicking the player.

* Check name length when setmyname is used

Added server-side check to prevent oversized names
when using the setmyname command. Oversized names are
truncated to 20 characters.

* Chat annoyance mitigation

The newline tag allows annoying impersonation.
Prevent this by adding "* " immediately after the newline.
This should be done server side so that it doesn't depend on
how the client handles the tag.
Messages containing these tags could be dropped, or the tag could be removed,
but this is a clean and simple way, and should be enough.

* Reverted 73fc318/59b009a

Removed newline tag filtering - it doesn't work well as the tag is case-insensitive
and can contain extra junk.

* Removed dead UDP master server

Removed dead UDP master server from the default configuration.

* Chat message length check

Check chat message length if configured to do so
- excessively long messages can cause lag and
annoyance as OLX seems to allow sending arbitrarily long messages.
Max length is 1024 by default - it should be reasonable.
Oversized messages are truncated. The sender can also be kicked but
this is off by default.

* Reduced logging

Previously, OLX logged all chat messages to the main log when hosting a server (dedicated or normal).
This cluttered the log and didn't provide very useful information.
Now this logging can be turned off.
NOTE that this has nothing to do with the "Log conversations" option - it goes to a different log!

* Fixed some -Wnarrowing warnings

Fixed some -Wnarrowing warnings.
Basically, picked up fixes from 22cb2696781ad628598f2da69f6afd86cd903b79
made by AMDmi3.

* [Dedicated/auxiliary]Update 0.58 pwn0meter.py to the 0.59 version

The old 0.58 version of this script contains bugs
so replace it with the 0.59 version that also has more features.